requirements.txt
matplotlib==3.3.1
pandas==1.3.5
Pillow==9.1.1
scikit_learn==1.1.1
seaborn==0.11.1
shap==0.40.0
streamlit==1.8.1

